refactor: extract `errorHandler` logic
